\name{HIV.db-package}
\alias{HIV.db-package}
\alias{HIV.db}
\docType{package}
\title{
HIV Annotation

}
\description{
provide the annotation data of HIV Genome and gene products and simple query interface 
The coordinates are based on HXB2.
}
\details{
\tabular{ll}{
Package: \tab HIV.db\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
License: \tab Artistic-2.0\cr
Depends: \tab R(>= 2.14.0)\cr
Imports: \tab Biobase (>= 2.5.5), methods,Biostrings\cr
biocViews: \tab Microarray, Preprocessing, Statistics\cr
URL: \tab http://www.rglab.org\cr
Built: \tab R 2.15.0; ; 2011-12-19 22:31:29 UTC; unix\cr
}

}
\author{
Mike Jiang,Raphael Gottardo

Maintainer: Mike Jiang <wjiang2@fhcrc.org>

}
\references{
\url{http://www.rglab.org/}
}
\keyword{ package }

