\name{readAlign}
\alias{readAlign}

\title{
Read multiple alignments in fasta format.
}

\description{
The function reads an alignment and returns a reference scale (refScale) and a reference sequence that can be used in loadFeatures to change the coordinate system.
}

\seealso{
\code{\link[Pviz]{readAlign}}
}
